Refactors scripts directory structure for clarity
