+++ scripts/data/importing/backfill_anbefaling_details.py
@@ -18,7 +18,7 @@ from pathlib import Path
 from typing import Dict, Any
 
 # Add project root to path
-project_root = Path(__file__).parent.parent.parent
+project_root = Path(__file__).parent.parent.parent.parent
 sys.path.insert(0, str(project_root))
