+++ scripts/data/maintenance/reduce_content.py
@@ -13,7 +13,7 @@ import json
 import sys
 from pathlib import Path
 
-project_root = Path(__file__).parent.parent.parent
+project_root = Path(__file__).parent.parent.parent.parent
 sys.path.insert(0, str(project_root))
