+++ scripts/ml/3_generate_embeddings.py
