+++ scripts/test/data/test_db_connection.py
@@ -2,7 +2,7 @@
 """Quick test to verify database connection and content."""
 import sys
 import os
-sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
+sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))))
 
 from app.config import settings
 import mysql.connector
